+++ src/metadata/index.ts
@@ -1,0 +1,2 @@
+export * from "./white-label.js";
+export * from "./provider.js";